feat: Add delete icon and card style parameter (#38)
* feat: Add delete icon card parameter

* feat: Add card style parameter

+++ siren-sdk/src/main/java/com/keyvalue/siren/androidsdk/core/elements/NotificationCard.kt
@@ -85,6 +85,7 @@ fun NotificationCard(
                     width = notificationCardStyle?.borderWidth!!,
                     color = notificationCardStyle.borderColor!!,
                 )
+                .then(cardProps?.cardStyle ?: Modifier)
                 .semantics { contentDescription = "siren-notification-card-${notification?.id}" },
     ) {
         val borderStroke =
@@ -211,17 +212,23 @@ fun NotificationCard(
             }
 
             if (cardProps?.hideDelete != true) {
-                Box(modifier = Modifier.weight(1f).padding(end = 6.dp), contentAlignment = Alignment.CenterEnd) {
-                    Icon(
-                        imageVector = Icons.Default.Clear,
-                        contentDescription = "siren-notification-delete-${notification?.id}",
-                        tint = themeColors?.deleteIcon!!,
-                        modifier =
-                            Modifier
-                                .size(notificationCardStyle.deleteIconSize!!)
-                                .clickable { deleteNotificationCallback() }
-                                .semantics { contentDescription = "siren-notification-delete-${notification?.id}" },
-                    )
+                if (cardProps?.deleteIcon != null) {
+                    cardProps.deleteIcon.also {
+                        it()
+                    }
+                } else {
+                    Box(modifier = Modifier.weight(1f).padding(end = 6.dp), contentAlignment = Alignment.CenterEnd) {
+                        Icon(
+                            imageVector = Icons.Default.Clear,
+                            contentDescription = "siren-notification-delete-${notification?.id}",
+                            tint = themeColors?.deleteIcon!!,
+                            modifier =
+                                Modifier
+                                    .size(notificationCardStyle.deleteIconSize!!)
+                                    .clickable { deleteNotificationCallback() }
+                                    .semantics { contentDescription = "siren-notification-delete-${notification?.id}" },
+                        )
+                    }
                 }
             }
         }
